refactor: move access utils since they are used by apps command
