+++ packages/cli/src/commands/apps/transfer.ts
@@ -4,11 +4,11 @@ import {Args, ux} from '@oclif/core'
 import * as Heroku from '@heroku-cli/schema'
 import {sortBy} from 'lodash'
 import * as inquirer from 'inquirer'
-import {getOwner} from '../../lib/access/utils'
+import {getOwner, isTeamApp, isValidEmail} from '../../lib/teamUtils'
 import lock from './lock'
 
-let AppTransfer = require('../../lib/app_transfer')
-let lock = require('./lock.js')[0]
+const AppTransfer = require('../../lib/app_transfer')
+
 function getAppsToTransfer(apps) {
   return inquirer.prompt([{
     type: 'checkbox', name: 'choices', pageSize: 20, message: 'Select applications you would like to transfer', choices: apps.map(function (app) {
@@ -33,18 +33,27 @@ export default class AppsTransfer extends Command {
       recipient: Args.string({description: 'user or team to transfer applications to', required: true}),
     };
 
+    static example = `$ heroku apps:transfer collaborator@example.com
+Transferring example to collaborator@example.com... done
+
+$ heroku apps:transfer acme-widgets
+Transferring example to acme-widgets... done
+
+$ heroku apps:transfer --bulk acme-widgets
+...`
+
     public async run(): Promise<void> {
       const {flags, argv, args} = await this.parse(AppsTransfer)
       const {app, bulk, locked} = flags
-      let recipient = args.recipient
+      const recipient = args.recipient
       if (bulk) {
-        let allApps = await this.heroku.get('/apps')
-        let selectedApps = await getAppsToTransfer(sortBy(allApps, 'name'))
+        const allApps = await this.heroku.get('/apps')
+        const selectedApps = await getAppsToTransfer(sortBy(allApps, 'name'))
         console.error(`Transferring applications to ${color.magenta(recipient)}...\n`)
-        for (let app of selectedApps.choices) {
+        for (const app of selectedApps.choices) {
           try {
-            let appTransfer = new AppTransfer({
-              heroku: heroku, appName: app.name, recipient: recipient, personalToPersonal: Utils.isValidEmail(recipient) && !Utils.isteamApp(app.owner), bulk: true,
+            const appTransfer = new AppTransfer({
+              heroku: this.heroku, appName: app.name, recipient: recipient, personalToPersonal: isValidEmail(recipient) && !isTeamApp(app.owner), bulk: true,
             })
             await appTransfer.start()
           } catch (error) {
@@ -52,13 +61,13 @@ export default class AppsTransfer extends Command {
           }
         }
       } else {
-        let appInfo = await this.heroku.get(`/apps/${app}`)
-        if (Utils.isValidEmail(recipient) && Utils.isteamApp(appInfo.owner.email)) {
+        const appInfo = await this.heroku.get(`/apps/${app}`)
+        if (isValidEmail(recipient) && isTeamApp(appInfo.owner.email)) {
           await cli.confirmApp(app, confirm, 'All collaborators will be removed from this app')
         }
 
-        let appTransfer = new AppTransfer({
-          heroku: heroku, appName: appInfo.name, recipient: recipient, personalToPersonal: Utils.isValidEmail(recipient) && !Utils.isteamApp(appInfo.owner.email),
+        const appTransfer = new AppTransfer({
+          heroku: this.heroku, appName: appInfo.name, recipient: recipient, personalToPersonal: isValidEmail(recipient) && !isTeamApp(appInfo.owner.email),
         })
         await appTransfer.start()
         if (locked) {
